Update ma.js
Major changes to address several goals coming up a lot less or more frequently than 1 in 3 (from 10% to 55%). A few conflict values also had to be tweaked to adjust for it, but the main changes are in weighing the rolling to adjust the frequency. Strongly weighted are the 4 tile Awards (Landlords etc.) in order to make it much more likely that at least one of them will be included. Weights are optimized for a limit value of 3.